fix(components): create and take out BaseBackgroundWrap
